Simulation area changes and crash hunt
Tried to hunt down the crash issue. It looks like for some reason it is getting "access none" when trying to read the particle system. Maybe it's to do with the garbage collection?

Also decreased the area of the simulation and increased the number of particles by 1000. Now there should be 6000 particles in total